DPDK: Use generic CPU instructions and skip AVX512
This commits removes AVX512 support and enables generic CPU options for
the DPDK build. The go-p4pack CI builds on Azure VMs which support
AVX512. However, since a lot of development systems (with WSL2 or
virtualbox don't support this (see [1] and [2]), we need to skip this
when building DPDK on the CI.

[1] https://forums.virtualbox.org/viewtopic.php?t=86047
[2] https://forums.virtualbox.org/viewtopic.php?f=7&t=87131

Thanks to Kyle Mestery <mestery@mestery.com> for finding the solution!

Signed-off-by: stolsma <github@tolsma.net>